final/requirements.txt
pillow
pytest
tkinter(usually comes pre-insalled with python)
